requirements.txt
pyyaml
numpy
Cython
h5py
Pillow
six
scipy
opencv-python
matplotlib
tb-nightly
future
yacs
gdown
flake8
yapf
isort==4.3.21
imageio
torch